requirements.txt
# Web backend
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# Security
bcrypt>=4.1
python-jose[cryptography]==3.3.0

# Email validation (for pydantic EmailStr)
email-validator>=2.2

# ML / NLP
scikit-learn==1.5.2
numpy==2.1.2
joblib==1.4.2

# GUI
PyQt6==6.7.1
requests==2.32.3

# Document export
python-docx==1.1.2
openpyxl==3.1.5

# Tests
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2

# Build
pyinstaller==6.10.0
